+++ utils/sv_idx_viewer28_draw.py
@@ -393,90 +393,3 @@ def draw_indices_2D_wbg(context, args):
         print(traceback.format_exc())
 
 
-# -------------------- BLF TEXT VIEWER ------------------- #
-
-
-
-def draw_3dview_text(context, args):
-
-    context = bpy.context
-    region = context.region
-    region3d = context.space_data.region_3d
-
-    geom, settings = args
-
-    txt_color = settings['font_text_color']
-    bg_color = settings['background_color']
-    scale = settings['scale']
-    draw_bg = settings['draw_background']
-    anchor = settings['anchor']
-
-    font_id = 0
-    text_height = int(13.0 * scale)
-    blf.size(font_id, text_height, 72)  # should check prefs.dpi
-
-    region_mid_width = region.width / 2.0
-    region_mid_height = region.height / 2.0
-
-    # vars for projection
-    perspective_matrix = region3d.perspective_matrix.copy()
-
-    final_draw_data = {}
-    data_index_counter = 0
-
-    def draw_all_text_at_once(final_draw_data):
-
-        # build bg mesh and vcol data
-        full_bg_Verts = []
-        add_vert_list = full_bg_Verts.extend
-        
-        full_bg_colors = []
-        add_vcol = full_bg_colors.extend
-        for counter, (_, _, _, _, _, type_draw, pts) in final_draw_data.items():
-            col = settings[f'bg_{type_draw}_col']
-            add_vert_list(pts)
-            add_vcol((col,) * 6)
-
-        # draw background
-        shader = gpu.shader.from_builtin('2D_SMOOTH_COLOR')
-        batch = batch_for_shader(shader, 'TRIS', {"pos": full_bg_Verts, "color": full_bg_colors})
-        batch.draw(shader)
-
-        # draw text 
-        for counter, (index_str, pos_x, pos_y, txt_width, txt_height, type_draw, pts) in final_draw_data.items():
-            text_color = settings[f'numid_{type_draw}_col']
-            blf.color(font_id, *text_color)
-            blf.position(0, pos_x, pos_y, 0)
-            blf.draw(0, index_str)
-
-    def gather_index(index, vec, type_draw):
-
-        vec_4d = perspective_matrix @ vec.to_4d()
-        if vec_4d.w <= 0.0:
-            return
-
-        x = region_mid_width + region_mid_width * (vec_4d.x / vec_4d.w)
-        y = region_mid_height + region_mid_height * (vec_4d.y / vec_4d.w)
-
-        # ---- draw text ----
-        index_str = str(index)
-        txt_width, txt_height = blf.dimensions(0, index_str)
-
-        # blf.position(0, x - (txt_width / 2), y - (txt_height / 2), 0)
-        pos_x = x - (txt_width / 2)
-        pos_y = y - (txt_height / 2)
-        # blf.draw(0, index_str)
-        pts = generate_points_tris(txt_width, txt_height, x, y-1)
-        data_index_counter = len(final_draw_data)
-        final_draw_data[data_index_counter] = (index_str, pos_x, pos_y, txt_width, txt_height, type_draw, pts)
-
-    # blf.color(font_id, *vert_idx_color)
-    if geom.locations_data and geom.text_data:
-        for text_item, (idx, location) in zip(geom.text_data, geom.locations_data):
-            gather_index(text_item, location, 'verts')
-    else:
-        for vidx in geom.locations_data:
-            gather_index(vidx[0], vidx[1], 'verts')
-
-
-    draw_all_text_at_once(final_draw_data)
